test(ios): split the server tests by what they cover
MediaUploadServerTests.swift held the server's admission tests, the
upload route's tests, and the uploader's, plus the helpers they shared.
Each now sits with its type: the server tests move under LocalServer
with the relay's, the route and uploader tests stay under Media, and
the bind probe, multipart builder, and browser-origin helper they share
become a support file. Assertions are unchanged.
